refactor: new remoe api helpers
